feature: use a float instead of a double distance() for consistancy

+++ corvis/src/feature/descriptor/orb_descriptor.h
@@ -53,8 +53,8 @@ public:
 
     orb_descriptor(float x, float y, const tracker::image& image);
     static bool is_better(const float distance1, const float distance2) {return distance1 < distance2;}
-    static double distance(const orb_descriptor &a, const orb_descriptor &b);
-    double distance(float x, float y, const tracker::image& image) const {
+    static float distance(const orb_descriptor &a, const orb_descriptor &b);
+    float distance(float x, float y, const tracker::image& image) const {
         return orb_descriptor::distance(*this, orb_descriptor(x,y,image));
     }
