+++ corvis/src/feature/detector/fast_10.cpp
@@ -4683,7 +4683,7 @@ xy fast_detector_10::track(const Descriptor& descriptor, const tracker::image& i
          else
           continue;
 
-        double score = descriptor.distance(x, y, image);
+        float score = descriptor.distance(x, y, image);
         if(Descriptor::is_better(score,best.score)) {
             best.x = (float)x;
             best.y = (float)y;
